ctmap: Factor out entry into separate file
