+++ pkg/maps/ctmap/entry.go
@@ -1,0 +1,61 @@
+// Copyright 2016-2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ctmap
+
+import (
+	"fmt"
+	"unsafe"
+
+	"github.com/cilium/cilium/pkg/byteorder"
+)
+
+// CtEntry represents an entry in the connection tracking table.
+type CtEntry struct {
+	rx_packets uint64
+	rx_bytes   uint64
+	tx_packets uint64
+	tx_bytes   uint64
+	lifetime   uint32
+	flags      uint16
+	// revnat is in network byte order
+	revnat         uint16
+	tx_flags_seen  uint8
+	rx_flags_seen  uint8
+	src_sec_id     uint32
+	last_tx_report uint32
+	last_rx_report uint32
+}
+
+// GetValuePtr returns the unsafe.Pointer for s.
+func (c *CtEntry) GetValuePtr() unsafe.Pointer { return unsafe.Pointer(c) }
+
+// String returns the readable format
+func (c *CtEntry) String() string {
+	return fmt.Sprintf("expires=%d rx_packets=%d rx_bytes=%d tx_packets=%d tx_bytes=%d flags=%x revnat=%d src_sec_id=%d\n",
+		c.lifetime,
+		c.rx_packets,
+		c.rx_bytes,
+		c.tx_packets,
+		c.tx_bytes,
+		c.flags,
+		byteorder.NetworkToHost(c.revnat),
+		c.src_sec_id)
+}
+
+// CtEntryDump represents the key and value contained in the conntrack map.
+type CtEntryDump struct {
+	Key   CtKey
+	Value CtEntry
+}
